% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Title Plot accidents on a state map for a given year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer representing the state number.}

\item{year}{An integer representing the
year for which the accident data is to be plotted.}
}
\value{
A plot of the accident locations on a state map.
If there are no accidents to plot,
a message is displayed and \code{NULL} is returned invisibly.
}
\description{
This function plots the locations of accidents on
a state map for a specified year.
It reads the accident data, filters it by the given state number,
and plots the longitude and latitude of the accidents.
}
\examples{
\dontrun{
  # Example usage:
  fars_map_state(1, 2013)
}
}
