% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Title Read a CSV file and return a tibble.}
\source{
extdate/accident_year.csv.bz2
}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A csv file that may or may not exist to import.}
}
\value{
A tibble containing the data from the CSV file.
If the csv does not exist, an error message will be shown.
}
\description{
This function reads a CSV file and returns its contents as a tibble.
It checks if the file exists before attempting to read it.
}
\examples{
\dontrun{
data <- fars_read("accident_2013.csv.bz2")
head(data, 5)
}
}
