% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Title Read accident data for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of integers representing
the years for which the data files are to be read.}
}
\value{
A list of tibbles, each containing
the accident data for a specific year.
If a file for a year does not exist,
\code{NULL} is returned for that year.
}
\description{
This function reads accident data files
for multiple years and returns a list of tibbles.
Each tibble contains the month and year columns.
If a file for a particular year does not exist,
a warning is issued and \code{NULL} is returned for that year.
}
\examples{
\dontrun{
     fars_read_years(1999)
     fars_read_years(as.list(1999, 2000, 2001))
     fars_read_years(1999:2016)
}
}
