% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Title Summarize accident data by year and month.}
\usage{
fars_summarize_years(years)
}
\value{
A tibble summarizing the number of accidents by year and month
}
\description{
This function reads accident data for multiple years, combines the data,
and summarizes the number of accidents by year and month.
}
\details{
#' @param years A vector of integers representing the years
for which the data files are to be read and summarized.
}
\examples{
\dontrun{
  # Example usage:
  summary <- fars_summarize_years(c(2013, 2014, 2015))
  print(summary)
}
}
