% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{make_filename}
\alias{make_filename}
\title{Title Generate a filename for accident data}
\source{
extdate/accident_year.csv.bz2
}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer representing the year for which
the filename is to be generated.}
}
\value{
A character string representing the filename.
}
\description{
This function generates a filename for accident data based
on the given year. The filename follows the format
"accident_\if{html}{\out{<year>}}.csv.bz2"
}
\examples{
\dontrun{
  # Example usage:
  filename <- make_filename(2013)
  print(filename)  # Outputs: "accident_2013.csv.bz2"
}
}
